Update keyAndMessageObj.java
This version of keyAndMessageObj removes the need for a separate CaesarCipher class. 

It decrypts/encrypts in this Class now. It also does this differently, instead of creating a dictionary of alphabetical characters with keys numbered 0-25, it simply splits the message into an array of individual letter strings and casts them into corresponding ASCII Integers, adds the key to each int, and then casts the elements back into the new letters and uses a StringBuilder to construct them into a new word.